+++ app_backend/src/metta/app_backend/clients/base_client.py
@@ -1,11 +1,10 @@
-from pathlib import Path
 from typing import Any, Type, TypeVar
 
 import httpx
-import yaml
 from pydantic import BaseModel
 
 from metta.app_backend.server import WhoAmIResponse
+from metta.common.auth.tokens import get_machine_token
 from metta.common.util.collections import remove_none_values
 from metta.common.util.constants import PROD_STATS_SERVER_URI
 
@@ -15,33 +14,6 @@ ClientT = TypeVar("ClientT", bound="BaseAppBackendClient")
 
 class NotAuthenticatedError(Exception):
     pass
-
-
-def get_machine_token(stats_server_uri: str | None = None) -> str | None:
-    """Get machine token for the given stats server.
-
-    Args:
-        stats_server_uri: The stats server URI to get token for.
-                         If None, returns token from env var or legacy location.
-
-    Returns:
-        The machine token or None if not found.
-    """
-    yaml_file = Path.home() / ".metta" / "observatory_tokens.yaml"
-    if yaml_file.exists():
-        with open(yaml_file) as f:
-            tokens = yaml.safe_load(f) or {}
-        if isinstance(tokens, dict) and stats_server_uri in tokens:
-            token = tokens[stats_server_uri].strip()
-        else:
-            return None
-    else:
-        return None
-
-    if not token or token.lower() == "none":
-        return None
-
-    return token
 
 
 class BaseAppBackendClient:
